Fix errors from the master merge
