+++ apps/ewallet/mix.exs
@@ -43,6 +43,7 @@ defmodule EWallet.Mixfile do
       {:bamboo_smtp, "~> 1.4.0"},
       {:bodyguard, "~> 2.2"},
       {:bypass, "~> 1.0.0", only: [:test]},
+      {:csv, "~> 2.0.0"},
       {:decimal, "~> 1.0"},
       {:deferred_config, "~> 0.1.0"},
       {:ewallet_config, in_umbrella: true},
