+++ apps/ewallet/test/ewallet/exporters/csv_exporter_test.exs
@@ -15,6 +15,7 @@
 defmodule EWallet.CSVExporterTest do
   use EWallet.DBCase
   import Ecto.Query
+  import EWalletDB.Factory
   alias EWallet.CSVExporter
   alias EWallet.Web.V1.CSV.TransactionSerializer
   alias EWalletDB.{Export, Transaction}
